// src/generated/headless-client-v1.d.ts
export interface paths {
// ...
    "/api/headless/v1/messages/{chat_room_uuid}/chat_room_messages": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Fetch chat room messages
         * @description Fetch chat room messages
         */
        get: {
            parameters: {
                query?: {
                    /** @description Reference record ID to fetch messages from, by default it is the last message ID */
                    id?: number;
                    /** @description Number of messages to be fetched after the ID record, by default, it is 20. */
                    next_per_page?: number;
                    /** @description Number of messages to be fetched before the ID record, by default, it is 20. */
                    previous_per_page?: number;
                };
                header?: never;
                path: {
                    /** @description Chat room UUID */
                    chat_room_uuid: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Ok */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["chat_room_messages"];
                    };
                };
                /** @description Unauthorized */
                401: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": unknown;
                    };
                };
            };
        };
        put?: never;
        /** Create chat room message */
        post: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    chat_room_uuid: string;
                };
                cookie?: never;
            };
            requestBody?: {
                content: {
                    "application/json": {
                        parent_message_id?: number | null;
                        rich_text_body: components["schemas"]["rich_text_body"];
                    };
                };
            };
            responses: {
                /** @description Created */
                202: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            creation_uuid?: string;
                            parent_message_id?: number | null;
                            sent_at?: string;
                        };
                    };
                };
                /** @description Unauthorized */
                401: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": unknown;
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        advanced_search_results: {
            count?: number;
            has_next_page?: boolean;
            page?: number;
            page_count?: number;
            per_page?: number;
            records?: (components["schemas"]["advanced_searched_post"] | components["schemas"]["advanced_searched_comment"] | components["schemas"]["advanced_searched_lesson"] | components["schemas"]["advanced_searched_space"] | components["schemas"]["advanced_searched_community_member"] | components["schemas"]["advanced_searched_group_mention"])[];
        };
// ...
        basic_post: {
// ...
            tiptap_body?: {
                attachments?: unknown[];
                body?: {
                    content?: Record<string, never>[];
                    type?: string;
                };
                circle_ios_fallback_text?: string;
                community_members?: components["schemas"]["shared_community_member"][];
                entities?: unknown[];
                format?: string;
                group_mentions?: unknown[];
                inline_attachments?: unknown[];
                polls?: unknown[];
                sgids_to_object_map?: Record<string, never>;
            } | null;
            topic_follower_id?: number | null;
            topics?: components["schemas"]["topic"][];
            unresolved_flagged_reports_count?: number;
            /** Format: date_time */
            updated_at?: string;
            url?: string;
            user_likes_count?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}